requirements.txt
pytest
pydantic
pyyaml
emoji
ttkbootstrap
reportlab
google-api-python-client

# Goocle Cloud Services SDKs
google-cloud-storage
google-cloud-bigquery
google-cloud-orchestration-airflow   # Composer
google-cloud-datacatalog
google-cloud-dataflow-client         # Dataflow (ojo: API limitada vía REST)
google-cloud-dataplex
google-cloud-dataproc
google-cloud-iam
google-cloud-logging
google-cloud-biglake
google-cloud-pubsub
google-cloud-resource-manager